feat: rewrite sd-drawer documentation
